refactor: clean model sources API
