+++ ras_commander/boundaries/__init__.py
@@ -1,0 +1,5 @@
+"""Boundary-condition data integrations."""
+
+from .coastal_boundary import CoastalBoundary
+
+__all__ = ["CoastalBoundary"]
